requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
requests==2.31.0
python-dotenv==1.0.0 